Optimise

+++ DoAnGame/FireBall.h
@@ -19,7 +19,7 @@
 
 class CFireBall : public CGameObject
 {
-	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects);
+	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
 	virtual void Render();
 public:
